+++ src/fr/adrienbrault/idea/symfony2plugin/asset/AssetGoToDeclarationHandler.java
@@ -8,7 +8,7 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiManager;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
 import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
 import org.apache.commons.lang.ArrayUtils;
 import org.jetbrains.annotations.Nullable;
 
@@ -34,7 +34,7 @@ public class AssetGoToDeclarationHandler implements GotoDeclarationHandler {
         }
 
         List<PsiElement> psiElements = new ArrayList<>();
-        for (VirtualFile virtualFile : TwigHelper.resolveAssetsFiles(psiElement.getProject(), psiElement.getText(), fileExtensionFilterIfValidTag)) {
+        for (VirtualFile virtualFile : TwigUtil.resolveAssetsFiles(psiElement.getProject(), psiElement.getText(), fileExtensionFilterIfValidTag)) {
             psiElements.add(PsiManager.getInstance(psiElement.getProject()).findFile(virtualFile));
         }
 
@@ -51,7 +51,7 @@ public class AssetGoToDeclarationHandler implements GotoDeclarationHandler {
 
         // asset / absolute_url dont have pre filter
         if(TwigPattern.getPrintBlockOrTagFunctionPattern("asset", "absolute_url").accepts(psiElement)) {
-            return (String[]) ArrayUtils.addAll(TwigHelper.CSS_FILES_EXTENSIONS, TwigHelper.JS_FILES_EXTENSIONS);
+            return (String[]) ArrayUtils.addAll(TwigUtil.CSS_FILES_EXTENSIONS, TwigUtil.JS_FILES_EXTENSIONS);
         }
 
         return null;
@@ -66,9 +66,9 @@ public class AssetGoToDeclarationHandler implements GotoDeclarationHandler {
 
             switch (tag) {
                 case "stylesheets":
-                    return TwigHelper.CSS_FILES_EXTENSIONS;
+                    return TwigUtil.CSS_FILES_EXTENSIONS;
                 case "javascripts":
-                    return TwigHelper.JS_FILES_EXTENSIONS;
+                    return TwigUtil.JS_FILES_EXTENSIONS;
                 default:
                     return null;
             }
