+++ src/fr/adrienbrault/idea/symfony2plugin/navigation/TwigFoldingBuilder.java
@@ -11,7 +11,6 @@ import com.jetbrains.twig.TwigFile;
 import fr.adrienbrault.idea.symfony2plugin.Settings;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
 import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.routing.Route;
 import fr.adrienbrault.idea.symfony2plugin.routing.RouteHelper;
 import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
@@ -57,7 +56,7 @@ public class TwigFoldingBuilder extends FoldingBuilderEx {
     private void attachTemplateFoldingDescriptors(PsiElement psiElement, List<FoldingDescriptor> descriptors) {
         // find path calls in file
         PsiElement[] fileReferences = PsiTreeUtil.collectElements(psiElement, psiElement1 ->
-            TwigPattern.getTemplateFileReferenceTagPattern().accepts(psiElement1) || TwigHelper.getFormThemeFileTag().accepts(psiElement1)
+            TwigPattern.getTemplateFileReferenceTagPattern().accepts(psiElement1) || TwigUtil.getFormThemeFileTag().accepts(psiElement1)
         );
 
         if(fileReferences.length == 0) {
