+++ src/fr/adrienbrault/idea/symfony2plugin/stubs/indexes/TwigExtendsStubIndex.java
@@ -12,7 +12,7 @@ import com.jetbrains.twig.TwigFile;
 import com.jetbrains.twig.TwigFileType;
 import com.jetbrains.twig.elements.TwigExtendsTag;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
 import gnu.trove.THashMap;
 import org.jetbrains.annotations.NotNull;
 
@@ -56,8 +56,8 @@ public class TwigExtendsStubIndex extends FileBasedIndexExtension<String, Void> 
             }
 
             for(PsiElement twigExtendsTag: twigExtendsTags) {
-                for (String s : TwigHelper.getTwigExtendsTagTemplates((TwigExtendsTag) twigExtendsTag)) {
-                    map.put(TwigHelper.normalizeTemplateName(s), null);
+                for (String s : TwigUtil.getTwigExtendsTagTemplates((TwigExtendsTag) twigExtendsTag)) {
+                    map.put(TwigUtil.normalizeTemplateName(s), null);
                 }
             }
