+++ src/fr/adrienbrault/idea/symfony2plugin/templating/inspection/TwigAssetsTagMissingInspection.java
@@ -6,7 +6,6 @@ import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
 import fr.adrienbrault.idea.symfony2plugin.templating.TwigPattern;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.templating.assets.TwigNamedAssetsServiceParser;
 import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
 import fr.adrienbrault.idea.symfony2plugin.util.service.ServiceXmlParserFactory;
@@ -38,12 +37,12 @@ public class TwigAssetsTagMissingInspection extends LocalInspectionTool {
                 if(TwigPattern.getAutocompletableAssetTag("stylesheets").accepts(element) && TwigUtil.isValidStringWithoutInterpolatedOrConcat(element)) {
 
                     String templateName = element.getText();
-                    if (!isKnownAssetFileOrFolder(element, templateName, TwigHelper.CSS_FILES_EXTENSIONS)) {
+                    if (!isKnownAssetFileOrFolder(element, templateName, TwigUtil.CSS_FILES_EXTENSIONS)) {
                         holder.registerProblem(element, "Missing asset");
                     }
                 } else if (TwigPattern.getAutocompletableAssetTag("javascripts").accepts(element) && TwigUtil.isValidStringWithoutInterpolatedOrConcat(element)) {
                     String templateName = element.getText();
-                    if (!isKnownAssetFileOrFolder(element, templateName, TwigHelper.JS_FILES_EXTENSIONS)) {
+                    if (!isKnownAssetFileOrFolder(element, templateName, TwigUtil.JS_FILES_EXTENSIONS)) {
                         holder.registerProblem(element, "Missing asset");
                     }
                 }
@@ -63,6 +62,6 @@ public class TwigAssetsTagMissingInspection extends LocalInspectionTool {
             }
         }
 
-        return TwigHelper.resolveAssetsFiles(element.getProject(), templateName, fileTypes).size() > 0;
+        return TwigUtil.resolveAssetsFiles(element.getProject(), templateName, fileTypes).size() > 0;
     }
 }
