+++ src/fr/adrienbrault/idea/symfony2plugin/twig/variable/collector/GlobalTwigConfigVariableCollector.java
@@ -9,9 +9,9 @@ import com.intellij.psi.util.CachedValue;
 import com.intellij.psi.util.CachedValueProvider;
 import com.intellij.psi.util.CachedValuesManager;
 import com.intellij.psi.util.PsiModificationTracker;
-import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.config.utils.ConfigUtil;
 import fr.adrienbrault.idea.symfony2plugin.stubs.ContainerCollectionResolver;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigUtil;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollector;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.TwigFileVariableCollectorParameter;
 import fr.adrienbrault.idea.symfony2plugin.templating.variable.dict.PsiVariable;
@@ -64,7 +64,7 @@ public class GlobalTwigConfigVariableCollector implements TwigFileVariableCollec
                 continue;
             }
 
-            for (Map.Entry<String, String> entry : TwigHelper.getTwigGlobalsFromYamlConfig((YAMLFile) file).entrySet()) {
+            for (Map.Entry<String, String> entry : TwigUtil.getTwigGlobalsFromYamlConfig((YAMLFile) file).entrySet()) {
                 String value = entry.getValue();
 
                 String serviceClass = ContainerCollectionResolver.resolveService(
